use context for shared state
